redis_service/requirements.txt
Flask==3.0.0
redis==5.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
pytest-mock==3.12.0
requests==2.31.0
